Settings + About + Fix Bug in Notifications Commit
Created 2 new Activities:
1. Settings - allow user to change the frequency in which he/she wants to grab alerts - TESTED
2. About - allow user to see desc of app and send a whatsapp message to technician with pre-defined text - TESTED

This is all viewable via the drawer temporarily placed in Alerts Activity.

BUG solved:
Notification to only get notification when there is an update, initially it grabbed the latest result when the activity started

POTENTIAL BUG:
Threading, when there are 2 users using the app, there will be 2 alerts triggered by the same update to the table cause they are using the same thread... not recreated.