some minor changes to ensure sysdiagnotif working
tested and working well with only 1 known bug, that sometimes the getnotificationsysdiag script will rerun when option is selected and intent is sent from settings to homeactivity. It is because there is a new instance of homeactivity is created when oncreating again. Bug cannot be recreated however a fail-safe method has been commented in this commit.